# docs/offeredapis/swagger/pms-api.yaml
openapi: 3.0.1
info:
  title: A1 Policy management service
  description: <h2>General</h2><p>The O-RAN Non-RT RIC Policy Management Service provides
    a REST API for management of A1 policices. <br/>The main tasks of the service
    are:</p><ul><li>A1 Policy creation, modification and deletion.</li><li>Monitoring
    and maintaining consistency of the SMO view of A1 policies and the Near-RT RICs</li><li>Maintaining
    a view of supported Near-RT RIC policy types</li><li>Supervision of using services
    (R-APPs). When a service is unavailble, its policies are removed.</li></ul><h2>APIs
    provided by the service</h2><h3>A1 Policy Management</h3><p>This is an API for
    management of A1 Policies.</p><ul><li>A1 Policy retrieval, creation, modification
    and deletion.</li><li>Retrieval of supported A1 Policy types for a Near-RT RIC</li><li>Retrieval
    of status for existing A1 policies</li></ul><h3>Management of configuration</h3><p>API
    for updating and retrieval of the component configuration. Note that there other
    ways to maintain the configuration.</p><h3>Callbacks</h3><p>These are endpoints
    that are invoked by this service. The callbacks are registerred in this service
    at service registration.</p><h3>NearRT-RIC Repository</h3><p>This is an API that
    provides support for looking up a NearRT-RIC. Each A1 policy is targeted for one
    Near-RT RIC.</p><h3>Health Check</h3><p>API used for supervision of the PMS component.</p><h3>Service
...
  license:
...
servers:
...
paths:
  /policy_types:
    get:
      tags:
      - A1 Policy Management V1.0
      summary: Query policy type names
      operationId: getPolicyTypes
      parameters:
      - name: ric
        in: query
        description: The name of the Near-RT RIC to get types for.
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        200:
          description: Policy type names
          content:
            '*/*':
              schema:
                type: array
                items:
                  type: string
        404:
          description: Near-RT RIC is not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/error_information'
  /a1-policy/v2/policy-instances:
    get:
      tags:
      - A1 Policy Management
      summary: Query for A1 policy instances
      description: Returns a list of A1 policies matching given search criteria. <br>If
        several query parameters are defined, the policies matching all conditions
        are returned.
      operationId: getPolicyInstances
      parameters:
      - name: policytype_id
        in: query
        description: The identity of the policy type to get policies for.
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: ric_id
        in: query
        description: The identity of the Near-RT RIC to get policies for.
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: service_id
        in: query
        description: The identity of the service to get policies for.
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        200:
          description: Policies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/policy_info_list_v2'
        404:
          description: Near-RT RIC, policy type or service not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_information'
...
  /a1-policy/v2/policy-types:
    get:
      tags:
      - A1 Policy Management
      summary: Query policy type identities
      operationId: getPolicyTypes_1
      parameters:
      - name: ric_id
        in: query
        description: The identity of the Near-RT RIC to get types for.
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        200:
          description: Policy type IDs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/policytype_id_list_v2'
        404:
          description: Near-RT RIC is not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_information'
  /a1-policy/v2/policies/{policy_id}:
    get:
      tags:
      - A1 Policy Management
      summary: Returns a policy
      operationId: getPolicy_1
      parameters:
      - name: policy_id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        200:
          description: Policy found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/policy_info_v2'
        404:
          description: Policy is not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_information'
    delete:
      tags:
      - A1 Policy Management
      summary: Delete a policy
      operationId: deletePolicy_1
      parameters:
      - name: policy_id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        200:
          description: Not used
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/void'
        423:
          description: Near-RT RIC is not operational
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/error_information'
        204:
          description: Policy deleted
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/void'
        404:
          description: Policy is not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/error_information'
...
  /policy:
    get:
      tags:
      - A1 Policy Management V1.0
      summary: Returns a policy configuration
      operationId: getPolicy
      parameters:
      - name: id
        in: query
        description: The identity of the policy instance.
        required: true
        style: form
        explode: true
        schema:
          type: string
      responses:
        200:
          description: Policy found
          content:
            '*/*':
              schema:
                type: object
        404:
          description: Policy is not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/error_information'
    put:
      tags:
      - A1 Policy Management V1.0
      summary: Put a policy
      operationId: putPolicy
      parameters:
      - name: type
        in: query
        description: The name of the policy type.
        required: false
        style: form
        explode: true
        schema:
          type: string
          default: ""
      - name: id
        in: query
        description: The identity of the policy instance.
        required: true
        style: form
        explode: true
        schema:
          type: string
      - name: ric
        in: query
        description: The name of the Near-RT RIC where the policy will be created.
        required: true
        style: form
        explode: true
        schema:
          type: string
      - name: service
        in: query
        description: The name of the service creating the policy.
        required: true
        style: form
        explode: true
        schema:
          type: string
      - name: transient
        in: query
        description: If the policy is transient or not (boolean defaulted to false).
          A policy is transient if it will be forgotten when the service needs to
          reconnect to the Near-RT RIC.
        required: false
        style: form
        explode: true
        schema:
          type: boolean
          default: false
      requestBody:
        content:
          application/json:
            schema:
              type: object
        required: true
      responses:
        200:
          description: Policy updated
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/void'
        201:
          description: Policy created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/void'
        423:
          description: Near-RT RIC is not operational
          content:
            '*/*':
              schema:
                type: string
        404:
          description: Near-RT RIC or policy type is not found
          content:
            '*/*':
              schema:
                type: string
    delete:
      tags:
      - A1 Policy Management V1.0
      summary: Delete a policy
      operationId: deletePolicy
      parameters:
      - name: id
        in: query
        description: The identity of the policy instance.
        required: true
        style: form
        explode: true
        schema:
          type: string
      responses:
        200:
          description: Not used
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/void'
        423:
          description: Near-RT RIC is not operational
          content:
            '*/*':
              schema:
                type: string
        204:
          description: Policy deleted
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/void'
        404:
          description: Policy is not found
          content:
            '*/*':
              schema:
                type: string
  /policies:
    get:
      tags:
      - A1 Policy Management V1.0
      summary: Query policies
      operationId: getPolicies
      parameters:
      - name: type
        in: query
        description: The name of the policy type to get policies for.
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: ric
        in: query
        description: The name of the Near-RT RIC to get policies for.
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: service
        in: query
        description: The name of the service to get policies for.
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        200:
          description: Policies
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/policy_info_v1'
        404:
          description: Near-RT RIC or type not found
          content:
            '*/*':
              schema:
                type: string
...
  /policy_status:
    get:
      tags:
      - A1 Policy Management V1.0
      summary: Returns a policy status
      operationId: getPolicyStatus
      parameters:
      - name: id
        in: query
        description: The identity of the policy.
        required: true
        style: form
        explode: true
        schema:
          type: string
      responses:
        200:
          description: Policy status
          content:
            '*/*':
              schema:
                type: object
        404:
          description: Policy is not found
          content:
            '*/*':
              schema:
                type: string
...
components:
  schemas:
    error_information:
      type: object
      properties:
        detail:
          type: string
          description: ' A human-readable explanation specific to this occurrence
            of the problem.'
          example: Policy type not found
        status:
          type: integer
          description: 'The HTTP status code generated by the origin server for this
            occurrence of the problem. '
          format: int32
          example: 503
      description: Problem as defined in https://tools.ietf.org/html/rfc7807
...